--- tests/model_hub_tests/pytorch/models/precommit
@@ -1,24 +1,0 @@
-facebook/opt-125m
-hf-tiny-model-private/tiny-random-CodeGenForCausalLM
-hf-tiny-model-private/tiny-random-GPT2LMHeadModel
-hf-tiny-model-private/tiny-random-OPTForCausalLM
-hf-tiny-model-private/tiny-random-GPTJForCausalLM
-hf-tiny-model-private/tiny-random-BloomForCausalLM
-hf-tiny-model-private/tiny-random-GPTNeoForCausalLM
-hf-tiny-model-private/tiny-random-GPTNeoXForCausalLM
-hf-internal-testing/tiny-random-GPTNeoForCausalLM
-hf-internal-testing/tiny-random-MptForCausalLM
-hf-internal-testing/tiny-random-BloomForCausalLM
-hf-internal-testing/tiny-random-GPTJForCausalLM
-hf-internal-testing/tiny-random-CohereForCausalLM
-hf-internal-testing/tiny-random-FalconForCausalLM
-hf-tiny-model-private/tiny-random-CodeGenForCausalLM
-hf-tiny-model-private/tiny-random-OPTForCausalLM
-hf-internal-testing/tiny-random-MistralForCausalLM
-hf-internal-testing/tiny-random-GPTNeoXForCausalLM
-hf-internal-testing/tiny-random-LlamaForCausalLM
-hf-internal-testing/tiny-random-StableLmForCausalLM
-hf-internal-testing/tiny-random-PhiForCausalLM
-hf-internal-testing/tiny-random-CodeGenForCausalLM
-hf-internal-testing/tiny-random-Starcoder2ForCausalLM
-hf-internal-testing/tiny-random-OPTForCausalLM
